ide: CLANG now default method in win32
git-svn-id: svn://ultimatepp.org/upp/trunk@14378 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Draw/ImageOp.cpp
@@ -1095,7 +1095,7 @@ Image GaussianBlur(const Image& img, int radius, bool co)
 	Vector<int> sizes;
 	Image t = img;
 	for (int i = 0; i < 3; i++)
-		t = ApplyBoxBlur(t, ((i < m ? wl : wu) - 1) / 2);
+		t = ApplyBoxBlur(t, int(((i < m ? wl : wu) - 1) / 2));
 	return pick(t);
 }
